+++ classes/event/run_aborted.php
@@ -1,0 +1,59 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Event: A run's pending attempts were aborted.
+ *
+ * @package    local_catquizlab
+ * @copyright  2026 Ralf Erlebach
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace local_catquizlab\event;
+
+/**
+ * A run's pending attempts were aborted.
+ */
+class run_aborted extends \core\event\base {
+    /**
+     * Initialise the event data.
+     *
+     * @return void
+     */
+    protected function init(): void {
+        $this->data['crud'] = 'u';
+        $this->data['edulevel'] = self::LEVEL_OTHER;
+        $this->data['objecttable'] = 'local_catquizlab_run';
+    }
+
+    /**
+     * The localised event name.
+     *
+     * @return string
+     */
+    public static function get_name(): string {
+        return get_string('event:run_aborted', 'local_catquizlab');
+    }
+
+    /**
+     * A human-readable description of what happened.
+     *
+     * @return string
+     */
+    public function get_description(): string {
+        return "The user with id '{$this->userid}' has aborted the pending attempts for run '{$this->objectid}'.";
+    }
+}